Add AffixButton to homepage to create new personality
