+++ src/components/Personality/PersonalityView.jsx
@@ -19,31 +19,10 @@ import { PlusOutlined } from "@ant-design/icons";
 import "./PersonalityView.css";
 import ReviewStats from "../ReviewStats";
 import { withTranslation } from "react-i18next";
+import AffixButton from "../Form/AffixButton";
 
 const { Title, Text, Paragraph } = Typography;
 const { Column } = Table;
-
-function AffixButton(props) {
-    // const [bottom, setBottom] = useState(10);
-    // @TODO use antd affix
-    return (
-        // <Affix offsetBottom={10}>
-        <Button
-            style={{
-                position: "fixed",
-                zInex: 9999,
-                bottom: "3%",
-                left: "85%"
-            }}
-            size="large"
-            shape="circle"
-            onClick={props.createClaim}
-            type="primary"
-            icon={<PlusOutlined />}
-        ></Button>
-        // </Affix>
-    );
-}
 
 function ClaimCard(props) {
     return (
